pyproject.toml
[tool.poetry]
name = "sergievsky"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
Pillow = "^9.1.1"
python-dotenv = "^0.20.0"
PyYAML = "^6.0"
django-ckeditor = "^6.4.2"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
